pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "readiness-assistant"
version = "1.0.0"
description = "AI Deployment Readiness Assistant MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.26.0",
    "pydantic",
    "requests",
    "google-generativeai",
    "sqlalchemy",
    "psycopg2-binary",
    "httpx",
    "python-dotenv",
    "pyyaml",
    # Async database drivers
    "asyncpg>=0.29.0",      # PostgreSQL async driver
    "aiomysql>=0.2.0",      # MySQL async driver
    "motor>=3.3.0"          # MongoDB async driver (async pymongo)
]

[project.scripts]
readiness-assistant = "server:run_server"
